refactor(vector): extract config/value types into engine-free vtypes leaf
