+++ vtypes/doc.go
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+// Package vtypes holds the pure, engine-free data types of the Rostam vector
+// API — the surface the wire codec and network client need without importing
+// the engine. The vector package re-exports every symbol here via aliases.
+package vtypes
